* Removed Untyped class: Call methods now return an object and you must cast it   to the correct type. * Added System.Object extension methods allowing us to chain calls. * Added System.Object extension method to do a cast which throws an exception   with a decent message if it fails (new --debug=casts switch doesn't work   on all platforms). * We no longer promote managed types to whatever the native calls expect: if the   native code expects an int then it's an error to pass anything else. * Type encoder no longer has the gross hack where small results were encoded   as "I" (Managed class now properly fills the return buffer for small types).

+++ source/CocoaException.cs
@@ -77,11 +77,11 @@ namespace MObjc
 					{
 						// If it does then get the serialized exception bytes,
 						IntPtr ptr = (IntPtr) data.Call("bytes");
-						int bytes = (int) data.Call("length");
+						uint bytes = (uint) data.Call("length");
 						
 						// copy them into a managed buffer,
 						byte[] buffer = new byte[bytes];
-						Marshal.Copy(ptr, buffer, 0, bytes);
+						Marshal.Copy(ptr, buffer, 0, (int) bytes);
 	
 						// and raise the original exception.
 						MemoryStream stream = new MemoryStream(buffer);
